[#116] refactor: 투표모달 타이머와 찬반 버튼만 리렌더링 되도록 수정
